staging client-side encrypted backup config
Uncomment to enable client-side encryption via aes-256-cbc with
passphrase.

Disabled for now - will uncomment on next rebuild or during maintenance
window given that the postgrescluster instance must be recreated in
order to put these settings in place without error. This setting is
immutable and the postgres operator complains in the logs when you
enable it on an existing cluster.